Refactor into DisassemblyHandler
