+++ src/main/java/gregicadditions/CommonProxy.java
@@ -195,7 +195,7 @@ public class CommonProxy {
 
     @SubscribeEvent(priority = EventPriority.LOWEST)
     public static void registerRecipesLowest(RegistryEvent.Register<IRecipe> event) {
-        RecipeHandler.buildDisassemblerRecipes();
+        DisassemblyHandler.buildDisassemblerRecipes();
         RecipeHandler.runRecipeGeneration();
     }
